Redid how parsing and instructions work.
The parse line functions now only grabs the label, instruction, and everything after the instruction.

The instruction validation is now separate from the actual execution. An instruction method (A method of all caps) now accepts a string of the text after the instruction and needs to derive its own parameters. These parameters are then validated and a function of the operation is returned. This function takes no arguments and performs the instruction.

Important note, a lambda function cannot be used here because lambda functions cannot contain assignments, only expressions.