Fixes #9, add test for this case, rename labels to be more descriptive
In order to cover all cases, the check was put into the closed function.
This means that an exception can be raied in execution that is not a
Hard Fault.

This was chosen because there may be a case where someone would want to
evaluate a code block where there was a load with a variable, but has
not yet defied that variable. The current function prints a warning
saying that it does not exist and that it should be declared before
running.

Two test cases were acced to check for this, one where everything is in
one code block, another where the variable are declared later.

the variable `label` was being used in different cases, this was
confusing and was fixed.